# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Reality Compiler API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List all orders
      responses:
        "200":
          description: Orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OrderSummary"
    post:
      operationId: placeOrder
      tags: [orders]
      summary: Place a new manufacturing order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlaceOrderInput"
      responses:
        "201":
          description: |
            Order draft created. If Stripe is configured the response includes
            a `checkoutUrl` the buyer must visit to actually pay; the order
            is not advanced past `pending_payment` until the
            `checkout.session.completed` webhook fires. When Stripe is not
            configured (dev / offline), the embedded `order` is returned
            directly with `paymentStatus = paid`.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlaceOrderResponse"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Quote not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get a single order with full detail
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /designer/orders:
    get:
      operationId: listDesignerOrders
      tags: [orders]
      summary: List orders placed against the authenticated user's published designs
      responses:
        "200":
          description: Designer orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DesignerOrderSummary"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /orders/{id}/advance:
    post:
      operationId: advanceOrder
      tags: [orders]
      summary: Advance the order to the next manufacturing status
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Already at terminal state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Supplier:
      type: object
      properties:
        id:
          type: integer
        slug:
          type: string
        name:
          type: string
        tagline:
          type: string
        description:
          type: string
        location:
          type: string
        country:
          type: string
        capabilities:
          type: array
          items:
            type: string
        materials:
          type: array
          items:
            type: string
        certifications:
          type: array
          items:
            type: string
        leadTimeMinDays:
          type: integer
        leadTimeMaxDays:
          type: integer
        pricingMultiplier:
          type: number
        setupFee:
          type: number
        rating:
          type: number
        capacityLevel:
          type: string
          enum: [low, medium, high]
      required:
        - id
        - slug
        - name
        - tagline
        - description
        - location
        - country
        - capabilities
        - materials
        - certifications
        - leadTimeMinDays
        - leadTimeMaxDays
        - pricingMultiplier
        - setupFee
        - rating
        - capacityLevel
    ProcessBreakdownItem:
      type: object
      properties:
        process:
          type: string
        description:
          type: string
        cost:
          type: number
      required:
        - process
        - description
        - cost
    QuoteScoreFactors:
      type: object
      properties:
        processMatch:
          type: number
        materialMatch:
          type: number
        leadTime:
          type: number
        rating:
          type: number
        total:
          type: number
      required:
        - processMatch
        - materialMatch
        - leadTime
        - rating
        - total
    Quote:
      type: object
      properties:
        id:
          type: integer
        sessionId:
          type: integer
        supplier:
          $ref: "#/components/schemas/Supplier"
        unitCost:
          type: number
        setupFee:
          type: number
        totalCost:
          type: number
        leadTimeDays:
          type: integer
        processBreakdown:
          type: array
          items:
            $ref: "#/components/schemas/ProcessBreakdownItem"
        scoreFactors:
          $ref: "#/components/schemas/QuoteScoreFactors"
        rank:
          type: integer
        notes:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - sessionId
        - supplier
        - unitCost
        - setupFee
        - totalCost
        - leadTimeDays
        - processBreakdown
        - scoreFactors
        - rank
        - notes
        - createdAt
    ShippingAddress:
      type: object
      properties:
        recipient:
          type: string
        line1:
          type: string
        line2:
          type: ["string", "null"]
        city:
          type: string
        region:
          type: string
        postalCode:
          type: string
        country:
          type: string
      required:
        - recipient
        - line1
        - city
        - region
        - postalCode
        - country
    PlaceOrderInput:
      type: object
      properties:
        quoteId:
          type: integer
        quantity:
          type: integer
          minimum: 1
        shippingAddress:
          $ref: "#/components/schemas/ShippingAddress"
        marketplaceListingId:
          type: ["integer", "null"]
      required:
        - quoteId
        - quantity
        - shippingAddress
...
    OrderStatusEvent:
      type: object
      properties:
        status:
          type: string
          enum: [queued, in_production, quality_check, shipped, delivered]
        note:
          type: string
        at:
          type: string
          format: date-time
      required:
        - status
        - note
        - at
    PaymentStatus:
      type: string
      enum:
        - pending_payment
        - paid
        - failed
        - refunded
        - partially_refunded
    OrderSummary:
      type: object
      properties:
        id:
          type: integer
        sessionId:
          type: integer
        sessionTitle:
          type: string
        productName:
          type: ["string", "null"]
        supplierName:
          type: string
        status:
          type: string
          enum: [queued, in_production, quality_check, shipped, delivered]
        quantity:
          type: integer
        totalCost:
          type: number
        payoutAmount:
          type: number
        designerUserId:
          type: ["string", "null"]
        leadTimeDays:
          type: integer
        paymentStatus:
          $ref: "#/components/schemas/PaymentStatus"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - sessionId
        - sessionTitle
        - supplierName
        - status
        - quantity
        - totalCost
        - payoutAmount
        - leadTimeDays
        - paymentStatus
        - createdAt
        - updatedAt
    DesignerOrderSummary:
      type: object
      properties:
        id:
          type: integer
        sessionId:
          type: integer
        sessionTitle:
          type: string
        productName:
          type: ["string", "null"]
        listingId:
          type: integer
        listingTitle:
          type: string
        listingDeleted:
          type: boolean
        buyerHandle:
          type: ["string", "null"]
        supplierName:
          type: string
        status:
          type: string
          enum: [queued, in_production, quality_check, shipped, delivered]
        quantity:
          type: integer
        totalCost:
          type: number
        payoutAmount:
          type: number
        leadTimeDays:
          type: integer
        paymentStatus:
          $ref: "#/components/schemas/PaymentStatus"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - sessionId
        - sessionTitle
        - listingId
        - listingTitle
        - listingDeleted
        - supplierName
        - status
        - quantity
        - totalCost
        - payoutAmount
        - leadTimeDays
        - paymentStatus
        - createdAt
        - updatedAt
    Order:
      type: object
      properties:
        id:
          type: integer
        sessionId:
          type: integer
        sessionTitle:
          type: string
        productName:
          type: ["string", "null"]
        quote:
          $ref: "#/components/schemas/Quote"
        supplier:
          $ref: "#/components/schemas/Supplier"
        quantity:
          type: integer
        totalCost:
          type: number
        payoutAmount:
          type: number
        designerUserId:
          type: ["string", "null"]
        leadTimeDays:
          type: integer
        shippingAddress:
          $ref: "#/components/schemas/ShippingAddress"
        status:
          type: string
          enum: [queued, in_production, quality_check, shipped, delivered]
        statusHistory:
          type: array
          items:
            $ref: "#/components/schemas/OrderStatusEvent"
        paymentStatus:
          $ref: "#/components/schemas/PaymentStatus"
        refundedAmount:
          type: number
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - sessionId
        - sessionTitle
        - quote
        - supplier
        - quantity
        - totalCost
        - payoutAmount
        - leadTimeDays
        - shippingAddress
        - status
        - statusHistory
        - paymentStatus
        - refundedAmount
        - createdAt
        - updatedAt
    PlaceOrderResponse:
      type: object
      description: |
        Result of `POST /orders`. When Stripe is configured the buyer is
        redirected to `checkoutUrl` to complete payment; the order is not
        advanced past `pending_payment` until the
        `checkout.session.completed` webhook fires. In dev/offline mode
        (`STRIPE_SECRET_KEY` unset) the embedded `order` is returned
        directly with `paymentStatus = paid`.
      properties:
        orderId:
          type: integer
        checkoutUrl:
          type: ["string", "null"]
        requiresPayment:
          type: boolean
        order:
          oneOf:
            - $ref: "#/components/schemas/Order"
            - type: "null"
      required:
        - orderId
        - checkoutUrl
        - requiresPayment
...