# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Reality Compiler API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderStatus } from "./orderStatus";
import type { OrderStatusEvent } from "./orderStatusEvent";
import type { PaymentStatus } from "./paymentStatus";
import type { Quote } from "./quote";
import type { ShippingAddress } from "./shippingAddress";
import type { Supplier } from "./supplier";

export interface Order {
  id: number;
  sessionId: number;
  sessionTitle: string;
  /** @nullable */
  productName?: string | null;
  quote: Quote;
  supplier: Supplier;
  quantity: number;
  totalCost: number;
  payoutAmount: number;
  /** @nullable */
  designerUserId?: string | null;
  leadTimeDays: number;
  shippingAddress: ShippingAddress;
  status: OrderStatus;
  statusHistory: OrderStatusEvent[];
  paymentStatus: PaymentStatus;
  refundedAmount: number;
  createdAt: Date;
  updatedAt: Date;
}