# lib/api-zod/src/generated/types/placeOrderResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Reality Compiler API
 * OpenAPI spec version: 0.1.0
 */
import type { Order } from "./order";

/**
 * Result of `POST /orders`. When Stripe is configured the buyer is
redirected to `checkoutUrl` to complete payment; the order is not
advanced past `pending_payment` until the
`checkout.session.completed` webhook fires. In dev/offline mode
(`STRIPE_SECRET_KEY` unset) the embedded `order` is returned
directly with `paymentStatus = paid`.

 */
export interface PlaceOrderResponse {
  orderId: number;
  /** @nullable */
  checkoutUrl: string | null;
  requiresPayment: boolean;
  order?: Order | null;
}